Move react query hooks into hooks folder
