+++ app/lib/api/reactQuery/hooks/query/maps.ts
@@ -1,7 +1,7 @@
 import { useQuery } from 'react-query';
-import queryClient from '../../reactQuery/queryClient';
-import QUERY_KEYS from '../../reactQuery/queryKeys';
-import API from '../../apiWrapper';
+import queryClient from '../../queryClient';
+import QUERY_KEYS from '../../queryKeys';
+import API from '../../../apiWrapper';
 
 export const useAllMaps = (searchString: string = '') => useQuery(
     QUERY_KEYS.allMaps(searchString),
